cambie los tipos de atributos de Ubicacion y saque el casteo del test

+++ proyecto-limpio-spring/src/main/java/ar/edu/unlam/tallerweb1/modelo/Ubicacion.java
@@ -11,26 +11,25 @@ public class Ubicacion {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	private Long longitud;
-	private Long latitud;
-	
-	
+	private double longitud;
+	private double latitud;
 	public Long getId() {
 		return id;
 	}
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public Long getLongitud() {
+	public double getLongitud() {
 		return longitud;
 	}
-	public void setLongitud(Long longitud) {
+	public void setLongitud(double longitud) {
 		this.longitud = longitud;
 	}
-	public Long getLatitud() {
+	public double getLatitud() {
 		return latitud;
 	}
-	public void setLatitud(Long latitud) {
+	public void setLatitud(double latitud) {
 		this.latitud = latitud;
 	}
+	
 }
